% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curvelo}
\alias{curvelo}
\title{ANEEL database on efficiency scores.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 61341 rows and 10 columns.}
\usage{
data("curvelo", package = "DTScanF")
}
\description{
List containing the eficiency for each company and a matrix indicating the neighbors.
}
\keyword{datasets}
