\name{DTScanF-package}
\alias{DTScanF-package}
\alias{DTScanF}
\docType{package}
\title{
\packageTitle{DTScanF}
}
\description{
\packageDescription{DTScanF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DTScanF}
\packageIndices{DTScanF}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DTScanF}

Maintainer: \packageMaintainer{DTScanF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
