% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_seqr2}
\alias{rcpp_seqr2}
\title{R2 Sequence}
\usage{
rcpp_seqr2(wind, window)
}
\arguments{
\item{wind}{A matrix to feed the regression.}

\item{window}{The size of the window}
}
\value{
a \code{list} containing:
\itemize{
  \item sequence: a \code{data.frame} indicating the starting and stoping 
  indexes and R2 for the calculated model.
  \item y: the target window. 
}
}
\description{
Search for the most similar time windows in a time series.
}
\author{
Leandro Mineti
}
