% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtsf.R
\name{dtsf}
\alias{dtsf}
\title{DTSF}
\usage{
dtsf(ts, poli, best, window, forecast)
}
\arguments{
\item{ts}{time series.}

\item{poli}{an integer defining the degree of the model used.}

\item{best}{an integer defining how many matches will be used in the forecasting.}

\item{window}{an integer indicating the size of the window to match.}

\item{forecast}{number of data points to be forecasted.}
}
\value{
\code{list}
}
\description{
Implementation a time series forecasting algorithm
}
\examples{

}
\author{
leandromineti@gmail.com
}
