% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtsf.R
\name{dtsf}
\alias{dtsf}
\title{DTSF}
\usage{
dtsf(ts, poli, best, window, forecast, reg = FALSE)
}
\arguments{
\item{ts}{time series.}

\item{poli}{an integer defining the degree of the model used.}

\item{best}{an integer defining how many matches will be used in the forecasting.}

\item{window}{an integer indicating the size of the window to match.}

\item{forecast}{number of data points to be forecasted.}

\item{reg}{logical indicating if series regularization should be performed. 
Default = FALSE.}
}
\value{
a \code{list} with two objects:
\itemize{
  \item sequence: a \code{data.frame} indicating the starting and stoping 
  indexes and R2 for the calculated model.
  \item forecast: a \code{matrix} of "best" rows and "forecast" columns.
}
}
\description{
Implementation a time series forecasting algorithm.
}
\examples{

}
\author{
leandromineti@gmail.com
}
