% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_seqr2}
\alias{rcpp_seqr2}
\title{Return seq}
\usage{
rcpp_seqr2(wind, window)
}
\arguments{
\item{wind}{A matrix to feed the regression.}

\item{window}{The size of the window}
}
\value{
a list
}
\description{
Return seq
}
\details{
Some description here
}
\author{
Leandro Mineti
}
